invert hierarchy of handle and state object
